merubah struktur, dependency
