+++ app/app.go
@@ -1,0 +1,20 @@
+package app
+
+import (
+	"github.com/gofiber/fiber/v2"
+	"github.com/muchlist/KalselDevApi/db"
+	"log"
+)
+
+func RunApp() {
+
+	// inisiasi database
+	client, ctx, cancel := db.Init()
+	defer client.Disconnect(ctx) //nolint:errcheck
+	defer cancel()
+
+	app := fiber.New()
+	mapUrls(app)
+	log.Fatal(app.Listen(":3500"))
+
+}
