+++ utils/mjwt/mjwt_test.go
@@ -6,18 +6,16 @@ import (
 	"time"
 )
 
-var (
-	Obj = NewJwt()
-)
-
 func TestJwtUtils_GenerateToken(t *testing.T) {
 	c := CustomClaim{
 		Identity:    "muchlis@gmail.com",
 		IsAdmin:     true,
 		ExtraMinute: 12,
+		Type:        Access,
+		Fresh:       true,
 	}
 
-	signedToken, err := Obj.GenerateToken(c)
+	signedToken, err := jwtObj.GenerateToken(c)
 
 	assert.Nil(t, err)
 	assert.NotEmpty(t, signedToken)
@@ -30,10 +28,10 @@ func TestJwtUtils_ValidateToken(t *testing.T) {
 		ExtraMinute: 12,
 	}
 
-	signedToken, err := Obj.GenerateToken(c)
+	signedToken, err := jwtObj.GenerateToken(c)
 	assert.Nil(t, err)
 
-	tokenValid, err := Obj.ValidateToken(signedToken)
+	tokenValid, err := jwtObj.ValidateToken(signedToken)
 
 	assert.Nil(t, err)
 	assert.NotEmpty(t, tokenValid)
@@ -42,7 +40,7 @@ func TestJwtUtils_ValidateToken(t *testing.T) {
 func TestJwtUtils_NotValidateToken(t *testing.T) {
 	invalidToken := "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9XX.eyJleHAiOjE2MDM4MDcyMzEsImlkZW50aXR5IjoibXVjaGxpc0BnbWFpbC5jb20iLCJpc19hZG1pbiI6dHJ1ZSwianRpIjoiIn0.dzKZdhPFtF-YC6uh5JZqBv7mhBjGTz1_rgIP-sRbYrU"
 
-	tokenValid, err := Obj.ValidateToken(invalidToken)
+	tokenValid, err := jwtObj.ValidateToken(invalidToken)
 
 	assert.Empty(t, tokenValid)
 	assert.NotNil(t, err)
@@ -56,10 +54,10 @@ func TestJwtUtils_ExpiredValidateToken(t *testing.T) {
 		ExtraMinute: -1,
 	}
 
-	signedToken, err := Obj.GenerateToken(c)
+	signedToken, err := jwtObj.GenerateToken(c)
 	assert.Nil(t, err)
 
-	tokenValid, err := Obj.ValidateToken(signedToken)
+	tokenValid, err := jwtObj.ValidateToken(signedToken)
 	assert.Nil(t, tokenValid)
 
 	assert.NotNil(t, err)
@@ -72,11 +70,11 @@ func TestJwtUtils_ReadToken(t *testing.T) {
 		IsAdmin:     true,
 		ExtraMinute: 0,
 	}
-	signedToken, err := Obj.GenerateToken(c)
+	signedToken, err := jwtObj.GenerateToken(c)
 	assert.Nil(t, err)
-	tokenValid, err := Obj.ValidateToken(signedToken)
+	tokenValid, err := jwtObj.ValidateToken(signedToken)
 	assert.Nil(t, err)
-	claims, err := Obj.ReadToken(tokenValid)
+	claims, err := jwtObj.ReadToken(tokenValid)
 	assert.Nil(t, err)
 
 	assert.Equal(t, "muchlis@gmail.com", claims.Identity)
